Move RE traits to dedicated locations.
