+++ resource_estimator/src/estimates.rs
@@ -6,16 +6,17 @@ pub use error::Error;
 mod error_budget;
 pub use error_budget::ErrorBudget;
 mod error_correction;
-pub use error_correction::{CodeWithThresholdAndDistance, CodeWithThresholdAndDistanceEvaluator};
+pub use error_correction::{
+    CodeWithThresholdAndDistance, CodeWithThresholdAndDistanceEvaluator, ErrorCorrection,
+};
 mod factory;
 pub use factory::{
-    BuilderDispatch2, DistillationRound, DistillationUnit, FactoryBuildError, FactoryDispatch2,
-    NoFactories, PhysicalQubitCalculation, RoundBasedFactory,
+    BuilderDispatch2, DistillationRound, DistillationUnit, Factory, FactoryBuildError,
+    FactoryBuilder, FactoryDispatch2, NoFactories, PhysicalQubitCalculation, RoundBasedFactory,
 };
 mod physical_estimation;
 pub use physical_estimation::{
-    ErrorCorrection, Factory, FactoryBuilder, FactoryPart, PhysicalResourceEstimation,
-    PhysicalResourceEstimationResult,
+    FactoryPart, PhysicalResourceEstimation, PhysicalResourceEstimationResult,
 };
 mod layout;
 mod logical_qubit;
